Changes:
1. Changed initialized params to float.
2. Added dictionary version of layer1 and above.

+++ frame_2D_alg/utils.py
@@ -94,7 +94,7 @@ def array2image(a):
 def imread(filename, raise_if_not_read=True):
     "Read an image in grayscale, return array."
     try:
-        return cv2.imread(filename, 0).astype(int)
+        return cv2.imread(filename, 0).astype(float)
     except AttributeError:
         if raise_if_not_read:
             raise SystemError('image is not read')
